11.4 Update 7
- extended HPL-2.3 building to support different BLAS_TYPE
- enhanced ethscpall to allow copying data with rsync, specifying net interface to use, and tar compression option
- changed FastFabric to setup PFC using software DCB (open lldp)
- renamed FastFabric component "Eth RDMA" to "Eth RoCE"
- added lldpad as prereq
- fixed ethlinkanalysis/ethfabricanalysis to honor the ConfigDir defined in mgt_config
- added rping loopback test and PFC mode check in hostverify
- improved hostverify to support passive mode intel_pstate
- fixed ethcabletest to make -n option works properly
- enhanced ethreport -A to include host ports that have no IP address
- cleaned up ethreport error messages
- enhanced ethbw to support duration argument
- updated man pages for FastFabric tools

+++ MpiApps/apps/get_mpi_blas.sh
@@ -41,19 +41,21 @@
 #
 # export MYVAR=`get_mpi_blas.sh`
 
-scratch=$(mktemp)
+if [[ -z ${BLAS_TYPE} ]]; then
+	scratch=$(mktemp)
+	rpm -qa >${scratch}
 
-rpm -qa >${scratch}
+	if [[ ! -z $MKLROOT && "$I_MPI_CC" == "icc" ]]; then 
+		# we only support MKL when compiling with Intel C.
+		export BLAS_TYPE="MKL"
+	elif grep -q openblas ${scratch}; then
+		export BLAS_TYPE="OPENBLAS"
+	else
+		export BLAS_TYPE="UNKNOWN_MPI_CC"
+	fi
 
-if grep -q mkl ${scratch} && which icc >/dev/null 2>&1; then 
-	# we only support MKL when compiling with Intel C.
-	export BLAS_TYPE="MKL"
-elif grep -q openblas ${scratch}; then
-	export BLAS_TYPE="OPENBLAS"
-else
-	export BLAS_TYPE="UNKNOWN_MPI_CC"
+	rm -f ${scratch}
 fi
 
-rm -f ${scratch}
 
 echo $BLAS_TYPE
